+++ src/table/batch/TableBatchOrchestrator.ts
@@ -132,6 +132,7 @@ export default class TableBatchOrchestrator {
 
     responseString += batchBoundary + "\r\n";
     // (currently static header) ToDo: Validate if we need to correct headers via tests
+    // We are doubling our change set boundary during errors here ??,--- --changesetresponse_b21a302d-b620-43ac-a8ba-c0f6e0445bcdchangesetresponse_b21a302d-b620-43ac-a8ba-c0f6e0445bcd\r\n
     responseString +=
       "Content-Type: multipart/mixed; boundary=" + changesetBoundary + "\r\n";
     changesetBoundary = "\r\n--" + changesetBoundary;
@@ -143,9 +144,11 @@ export default class TableBatchOrchestrator {
       });
     } else {
       // serialize the error
-      responseString += changesetBoundary;
+      responseString += changesetBoundary + "\r\n";
       // then headers
-      // content type etc
+      responseString += "Content-Type: application/http\r\n";
+      responseString += "Content-Transfer-Encoding: binary\r\n";
+      responseString += "\r\n";
       // then HTTP/1.1 404 etc
       responseString += this.errorResponse;
     }