+++ src/table/batch/TableBatchSerialization.ts
@@ -605,13 +605,8 @@ export class TableBatchSerialization extends BatchSerialization {
     contentID: number,
     request: BatchRequest
   ): string {
-    const changesetBoundary = this.changesetBoundary.replace(
-      "changeset",
-      "changesetresponse"
-    );
     let errorReponse = "";
     const odataError = err as StorageError;
-    errorReponse += changesetBoundary + "\r\n";
     // Errors in batch processing generate Bad Request error
     errorReponse = this.serializeHttpStatusCode(errorReponse, 400);
     errorReponse += "Content-ID: " + contentID + "\r\n";